[#235] :recycle: refactor: 새 스티커 편집 기능 custom hook으로 분리
